requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-ReCaptcha==2.3.1
redis==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.2.0